+++ src/layouts/Pane/Pane.layout.tsx
@@ -25,6 +25,7 @@ import {
   useHtmlAttribute,
   cssRemToCssPx,
   extractPx,
+  clamp,
 } from "@utils";
 
 import type { TestIdProps } from "@types";
@@ -41,16 +42,14 @@ const defaultPaneProps: PaneDefaultProps = {
   styleOverride: null,
   anchor: "left",
   top: 0,
-  bottom: 0,
+  bottom: "unset",
   expandedRatio: 0.5,
 };
 
 /**
- * Pane component
+ * An expandable & collapsible pane wrapper layout with configurable styles & behavior
  *
- * TODO description
- *
- * @version 0.0.1
+ * @version 0.0.2
  *
  * @param {PaneProps & TestIdProps} props - Pane component props
  * @returns {JSX.Element}
@@ -104,7 +103,10 @@ const Pane = forwardRef<HTMLElement, PaneProps & TestIdProps>(
       [expandedWidthPx, collapsedWidthPx]
     );
 
-    const margin = useMemo(() => diff * expandedRatio, [diff, expandedRatio]);
+    const margin = useMemo(
+      () => diff * clamp(expandedRatio, 0, 0.9999999999999999),
+      [diff, expandedRatio]
+    );
 
     const containerRef = useRef<Nullable<HTMLElement>>(null);