+++ CHANGELOG.md
@@ -10,6 +10,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - `v0.0.3` DateInput, `v0.0.7` MultiSelectInput, `v0.0.20` SelectInput, `v0.0.8` Input components
+  - Tweak styles to make most inputs take all available height in tables containing taller cells
+
+- `v0.0.2` SideEntityMenu component
+  - Update styles while expanded via breakpoint
+
+- `v0.0.2` Pane component
+  - Update styles, write tests & documentation
 
 ## 1.9.0 - 2025-07-09
 ### Added
@@ -17,7 +24,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - New side navigation menu component for entity-based navigation
 - `v0.0.1` Pane layout
   - Expandable & collapsible container
-  - Tweak styles to make most inputs take all available height in tables containing taller cells
 
 ## 1.8.32 - 2025-07-04
 ### Fixed